Update Persona Module
So, this is a revised and almost original code to Persona Module to RPG Maker MV.
The Persona List is working and you can define who actor is a persona put the note tag <tipo:Persona> in the character. I can't due this without the help of GabrielSchade from CtrlAlt Discord Group. Thanks, bro!